\name{intercrop}
\alias{intercrop}
\docType{data}
\title{
Intercropping data
}
\description{
This is a generated dataset containing of 4 different variables, measured across 40 plots on a strip-type intercropping design consisting of 2 crops.
}
\usage{data("intercrop")}
\format{
  The format is:
40 by 4 matrix
}
\details{
Contains generated data similar to the data used in the Hermes et al. (2023) paper, except that this data consists of a a single row of alternating crops.
}
\source{
Generated
}
\references{
1. Hermes, S., van Heerwaarden, J., and Behrouzi, P. (2023). A Spatial Autoregressive Graphical Model with Applications in Intercropping. arXiv preprint, arXiv:2308.04325. \cr
}
\examples{
data(intercrop)
}
\keyword{datasets}
